(wip) preserve existing dependency revisions
This changes the default behavior of 'godep save' to
preserve any dependencies already in the list with their
existing revision. New dependencies will be added, and
old dependencies will be removed, but that's all. This
makes the save command roughly analogous to 'go get'
with no flags.

PENDING:
We also introduce a new command, 'godep update', to
update one or more dependency versions. This is roughly
analogous to 'go get -u'. This is a separate command
(rather than 'godep save -u') because the package
argument is different: it specifies a dependency rather
than a top-level package.

fixes #65