godep save: add -tags flag.
This flag can be used to build the dependencies for a project
that uses build tags for conditional compilation.  For example,
`godep save -tags heroku` will produce dependencies for assuming
that the 'heroku' tag is set.

Because package names are passed using an arbitrary length names
arg in pkg.go, I opted to create a new function that accepts tags
in order to minimize the amount of refactoring that would be
needed.  It seems that `go list` works perfectly fine with the
tags flag.

All save tests pass, but I have not yet figured out how to modify
the contents of a file to make a valid test for this code.