godep version command
Output the version as well as some godep runtime information that is
useful for debugging user's issues.

The version const would be bumped each time a PR is merged into master
to ensure that we'll be able to tell which version someone got when they
did a `go get github.com/tools/godep`.

This is similar to what github.com/kr/okjson does.

Other suggestions welcome, but I need to introduce the idea of a version
for debugging purposes.