Vendor applications safely
For go 1.4, adding a dummy go file with some import clauses importing
"main" packages was enough for godep to vendor them. With go 1.5 it is
not working anymore, because go will bail out when loading a package
with this dummy go file with an error 'import 'some/main/package' is a
program, not an importable package'.

This introduces reading the "appVendor" file, which is basically a
list of main packages to import. It is not a go source file, so it
sidesteps the issue with the new go.

The reason to add this is that the project may want to use some
application from specific revision during building or testing. git
submodules allow that, but the problem can arise when the project
disappears. "go get" does not allow getting a specific revision.