Fix loading a dot package
Getting the dot package should succeed, even in spite of no buildable
Go sources in the directory. The only thing we need from the package
is the import path and that was provided even by `go list -e -json .`.

If the toplevel directory has no buildable Go sources and we invoke
godep save without any args, then it will fail later. For such
projects an arg like ./... should be passed.