Stow away the helper files when trying to import the package
There are many projects outside that have some helper binaries with a
"// +build ignore" clause and "main" package in a directory, where the
rest of files have some other package name.

In this commit we are renaming them, so they won't get picked up by
build context. After we are done, they are renamed back.