Ignore .go files when checking for license/legal
The current heuristic for license or legal files is a little broad
because it will inadvertently sweep up go source files that shouldn't be
vendored. As one example, when trying to vendor the package
`github.com/camlistore/go4/errorutil`, `godep save` also brings in the
file `github.com/camlistore/go4/blob/master/legal/legal.go`.

This patch changes the `IsLicenseFile` and `IsLegalFile` checks to
immediately return false if the filename ends in `.go`.