Use a fixed abbreviation size in `git describe` output
Currently, this setting depends on whatever the user of `godep` has in
their `git config --get core.abbrev`, if it is set.

Even when it is not set, the number of characters used in the shortened
git commit will depend on whether there are collisions for that prefix
on the local repository (which in large part depends on how much
activity exists there.)

As a result, when multiple users are maintaining Godeps/ for a single
project, many spurious changes to Godeps/Godeps.json are generated due
to the lack of stability in the length of the abbreviated commit in
output of `git describe`.

Let's fix this by enforcing a reasonable abbreviation length in godep's
use of `git describe`. 14 characters is very unlikely to result in
collisions for prefixes.